refactor: move GitHubUrl::find_current_location function to RepoManager
