+++ rook/src/git/repo.rs
@@ -1,6 +1,9 @@
 use std::{env, fs, path::PathBuf};
 
 use git2::Repository;
+use tracing::error;
+
+use crate::{GitHubUrl, file_exists_in_repo, find_last_commit_id, track_file_rename_in_commit};
 
 /// A guard that automatically removes a temporary directory when dropped.
 pub struct TempDirGuard {
@@ -32,6 +35,17 @@ pub struct RepoManager {
 }
 
 impl RepoManager {
+    /// Clones a Git repository from a GitHub URL.
+    ///
+    /// # Arguments
+    /// * `url` - The GitHub URL of the repository to clone
+    ///
+    /// # Returns
+    /// A `RepoManager` instance that will automatically clean up the cloned repository when dropped.
+    pub fn from_github_url(url: &GitHubUrl) -> Result<Self, git2::Error> {
+        Self::clone_repo(&url.clone_url(), url.branch())
+    }
+
     /// Clones a Git repository, optionally cloning only a specific branch.
     ///
     /// When a branch name is provided, only that specific branch will be cloned,
@@ -67,6 +81,52 @@ impl RepoManager {
             repo,
             _temp_dir_guard,
         })
+    }
+
+    /// Attempts to find the current location of a file in the repository
+    ///
+    /// # Returns
+    /// * `Ok(Some(String))` - The current location of the file if found
+    /// * `Ok(None)` - If the file was not found
+    /// * `Err(git2::Error)` - If there was an error accessing the repository
+    pub fn find_current_location(
+        &self,
+        github_url: &GitHubUrl,
+    ) -> Result<Option<String>, git2::Error> {
+        let file_path = github_url
+            .file_path()
+            .ok_or_else(|| git2::Error::from_str("No file path in URL"))?;
+
+        let repo = self.get_repo();
+        let mut current_path = file_path.to_string();
+
+        loop {
+            if file_exists_in_repo(repo, &current_path)? {
+                return Ok(Some(current_path));
+            }
+
+            let commit = match find_last_commit_id(&current_path, repo) {
+                Ok(commit) => commit,
+                Err(e) => {
+                    error!("Error finding last commit for {}: {}", current_path, e);
+                    return Ok(None);
+                }
+            };
+
+            match track_file_rename_in_commit(repo, &commit, &current_path)? {
+                Some(new_path) => {
+                    current_path = new_path;
+                }
+                None => {
+                    error!(
+                        "Could not find new path for {} in commit {}",
+                        current_path,
+                        commit.id()
+                    );
+                    return Ok(None);
+                }
+            }
+        }
     }
 
     /// Returns a reference to the managed Git repository.
@@ -124,5 +184,49 @@ mod tests {
 
         assert!(repo_manager.get_repo().head().is_ok());
         assert!(repo_manager.get_repo().head().unwrap().name().unwrap() == "refs/heads/maout");
+    }
+
+    #[test]
+    /// This test is related to `file_tracker::tests::test_track_file_rename_in_commit_with_multiple_moves`
+    /// which demonstrates the same file movement pattern:
+    /// 1. Initially located at: tmp.txt (root directory)
+    /// 2. First moved to: dockerfile_history/tmp.txt
+    /// 3. Finally moved to: img/tmp.txt
+    fn test_find_current_location() {
+        let url = GitHubUrl::parse(
+            "https://github.com/reddevilmidzy/zero2prod/blob/test_for_queensac/tmp.txt",
+        )
+        .unwrap();
+
+        let repo_manager = RepoManager::from_github_url(&url).unwrap();
+
+        assert_eq!(
+            repo_manager.find_current_location(&url).unwrap(),
+            Some("img/tmp.txt".to_string())
+        );
+    }
+
+    #[test]
+    /// This test verifies the behavior when a file cannot be found in the repository.
+    /// It tests two scenarios:
+    /// 1. A file that never existed in the repository
+    /// 2. A file that was deleted and not moved anywhere else
+    fn test_find_current_location_file_not_found() {
+        // Test case 1: File that never existed
+        let url =
+            GitHubUrl::parse("https://github.com/reddevilmidzy/kingsac/blob/main/non_existent.txt")
+                .unwrap();
+
+        let repo_manager = RepoManager::from_github_url(&url).unwrap();
+        assert_eq!(repo_manager.find_current_location(&url).unwrap(), None);
+
+        // Test case 2: File that was deleted
+        let url = GitHubUrl::parse(
+            "https://github.com/reddevilmidzy/kingsac/blob/main/will_be_deleted.rs",
+        )
+        .unwrap();
+
+        let repo_manager = RepoManager::from_github_url(&url).unwrap();
+        assert_eq!(repo_manager.find_current_location(&url).unwrap(), None);
     }
 }
